Re-format all C/C++ code with clang-format-14

+++ core/kernel_linux/kvm_hypercall.h
@@ -11,18 +11,16 @@
 
    The return value is placed in the guest's RAX register.
 */
-typedef unsigned long(*kvm_hypercall_callback_t)(
-                struct kvm_vcpu* vcpu,
-                unsigned long nr,
-                unsigned long a0,
-                unsigned long a1,
-                unsigned long a2,
-                unsigned long a3); 
+typedef unsigned long (*kvm_hypercall_callback_t)(struct kvm_vcpu *vcpu, unsigned long nr,
+                                                  unsigned long a0, unsigned long a1,
+                                                  unsigned long a2, unsigned long a3);
 
 /* These functions are defined in arch/x86/kvm/x86.c. The function's return
  * value is placed in the guest's RAX register.
  */
-void kvm_register_hypercall_callback(kvm_hypercall_callback_t callback);
-void kvm_remove_hypercall_callback(kvm_hypercall_callback_t callback);
+void
+kvm_register_hypercall_callback(kvm_hypercall_callback_t callback);
+void
+kvm_remove_hypercall_callback(kvm_hypercall_callback_t callback);
 
 #endif
